Changed board index method and improved ConveyorBelt register action
* Made the board index method accept plot-style coordinates instead of
matrix-style coordinates. The method converts to matrix-style
coordinates and uses these to index the matrix.
* Removed the board index method from the IBoard interface in order to
fully encapsulate the Cell helper class. In line with this, also made
Cell a package-scope class within the Board file.
* The performRegisterAction method in ConveyorBelt now checks if there
is something blocking the way of the robot's movement. If there is, it
does not perform the movement.
* Renamed ConveyerBelt to the proper spelling, ConveyorBelt. Eclipse's
spell checker might complain, but it's wrong.
* Renamed moveX and moveY in the Piece class to the less confusing names
shiftX and shiftY.
* Added a coordinateWithinBounds method to the IBoard interface with an
implementation in Board, as we will probably need it later.