Wrote ConveyorBelt tests, refactored things, and implemented things
- Implemented Pieces updating the board matrix when they move. This was
done using an updatePiecePosition method in Board and IBoard, which the
pieces call whenever they move
- Added placeRobot and placeEElement methods in Board and IBoard
- Implemented four basic conveyor belt tests using test context and
dependency injection. The tests all succeeded (when the other step
definition files with compilation errors were taken out). In connection
with this,
	- Started using domain-specific step definition files instead of
feature-specific ones
	- Started using a new syntax for step definition files, where they are
prefixed with "SD" instead of "StepDefinition" for brevity
- Added a constructor for Board which initializes it with a matrix of
empty Cells
- Moved Player class to a new package called player