Reimplemented BoardPanel using paintComponent
- Deleted ImagePanel
- Moved the scaledImage method to a new utility class, ImageUtils. Also
added a smaller rawImage method which is not being used anywhere at the
moment
- Added an ISprite interface and an implementing Sprite class. Each
sprite has an associated image, as well as x and y coordinates and
degrees of rotation. ISprite contains a single method, drawUsing, which
takes a Graphics2D object and draws the sprite itself using that
Graphics2D instance. This was implemented in the Sprite class using the
AffineTransform class from the Java library, which does linear algebra
on an image. AffineTransform was not needed for drawing the image at a
specific location, but it was needed for rotating the image, so I just
decided to use it for both purposes
- Added a SpriteFactory class with a method called getFromPieceID. This
takes a string, eventually planned to be the piece ID of a piece on the
board, and tries to find an image in the images folder that has that ID
as its filename. It then returns a new Sprite with this image
- BoardPanel has a list of ISprites. The paintComponent method now
paints the background pattern, and then iterates through all its sprites
and calls drawUsing on them with the Graphics2D object