Fixed an edge case related to respawn points
- Fixed the case where a robot dies on its own respawn point. It used to
enter an infinite recursion, because it would see itself on the respawn
point and kill itself to make room. Now, it only kills the robot on the
respawn point if that robot is not itself
- Also had to fix the moveRobotFromTo method in Board (and
moveEElementFromTo) in the case where the old position and new position
are the same. It used to nullify the robot at that position without
replacing it, and also call performImmediateAction on the EElement at
that position. It now only tries to move the robot/EElement if the
positions are different