Commented out animations for now, wrote ChainingPanel sprite change
- Added an extra check to the degDiff logic, for the case where degCurr
is 0 and degNew is 270
- Commented out the animations because they currently suffer from a
nondeterministic thread bug. Basically, if you rotate a robot twice, the
animation will look weird, but even worse, there is no guarantee that
the first animation will finish before the second. So the second may
snap the sprite before the first, leading to a wrong final sprite
orientation
- Added a new activate() method to Sprite, which does nothing
- Created a new class ImageToggleSprite which subclasses Sprite, and
overrides the activate() method to switch its image between a default
image and an alternative image
- SpriteFactory now checks for ChainingPanels, and if it is given one,
it returns an ImageToggleSprite toggling between the two chaining panel
sprites
- When BoardPanel receives an activation event, it activates the sprite
at the position given by the event
- Added new movement firePropertyChanges to Robot to work with Arooj's
changes