New complex queued animation implementation, fixing the previous bug
- Created a new package called animations

Created a new abstract class called Animation:
- The constructor takes a duration parameter in milliseconds and uses
this, along with an FPS of 60, to compute the number of frames needed
for an animation. This is then returned by the getNumFrames method
- There are three abstract methods: initialize(), establishNextFrame(),
and finalize(). These are used by MasterView as described below

MasterView:
- Now has a Queue of Animation called animationQueue, and a method
called enqueueAnimation, which adds an animation to this queue
- Added a method called playAllAnimations, which dequeues an animation
from animationQueue, calls initialize() on it, and creates a new thread
to play this animation. The thread loops "frames" times, where frames is
obtained by calling getNumFrames on the animation. In each loop, it
calls establishNextFrame(), and then sleeps for one sixtieth of a second
in order to maintain 60 FPS. Once the thread is finished, it calls
finalize() on the animation, and then calls playAllAnimations() again
recursively. The process repeats until the queue is empty. Also wrote a
documentation comment explaining the method

MovementEvent:
- Changed MovementEvent to contain the robot number of the robot that
needs to be moved, and a Position detailing the position difference
- Reimplemented the subtract method for Position. This is now used by
Robot when firing a MovementEvent, in order to calculate the position
change

Movement animations:
- Created a new class called SpriteMovementAnimation which extends
Animation, and copied most of the old movement animation code to this
class, splitting it between the initialize(), establishNextFrame(), and
finalize() methods
- In addition to the duration parameter needed for the super
constructor, the constructor of this class takes the Sprite it is
required to move, as well as two integers specifying the x and y
displacements of the movement

BoardPanel:
- In order to go from the robot number provided by MovementEvent to the
Sprite required by SpriteMovementAnimation, BoardPanel now has a HashMap
called robotNumToSpriteMap. This maps from robot numbers to the sprites
representing those robots
- This map is updated whenever a new Robot sprite is added, in the
addSprite method
- BoardPanel now has a reference to its MasterView, in order to enqueue
animations
- The moveRobot method now computes everything needed by a
SpriteMovementAnimation, and then enqueues one using these parameters

- Added a new property change event called ProgrammingPhaseBeginEvent.
For now, this contains nothing, but when MasterView receives one, it
calls playAllAnimations()
- Updated the demo to show the new animation playback feature
- Fixed the getNumRows and getNumColumns methods in Board