New sprites and bug fixes
- Fixed a bug in Board: When eElements are removed, the board now
correctly removes them from the piece list. If you try to remove an
element that does not exist, it now throws a NullPointerException
- Fixed a bug in Game: HealthStation.ID is now included in
registerActorPriorityList
- Renamed laser sprite to the laser piece ID. As a consequence,
SpriteFactory now works for lasers
- Added blue and orange teleporter sprites, and modified SpriteFactory
to give the orange one for receiving teleporters and the blue one for
sending teleporters. TODO: Make it so whether a teleporter is sending or
not is initialized in the constructor
- The demo now demonstrates board lasers and health stations