Advanced chain animations
-Oh boy will this need to be dealt with.
-Chains now have fancy animation.
-problem exposed with chaining movement, if a robot pulls on another
while it falls into a pit its movement was executed first de-chaining in
the process and then the other one tries to move, but because we refer
to it from the chaining, which is now null an exception occurs.
For now its fixed by pushing the pulled robot first and then doing our
movement.
-There will need to be done some recaftoring as I made two new classes
chainPair and Triplet for this animation to work.
-these two classes could be used for the laser animation as they are a
way better solution than the current one.